src/main/java/org/example/registration_with_check/service/Utility.java
package org.example.registration_with_check.service;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.nio.charset.StandardCharsets;
import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.util.Random;
import java.util.concurrent.TimeUnit;

public class Utility {

    private Utility(){

    }

    private static final Logger log = LoggerFactory.getLogger(Utility.class);

    public static String sha256HexString(String originalString) throws NoSuchAlgorithmException {
        MessageDigest digest = MessageDigest.getInstance("SHA-256");
        byte[] encodedhash = digest.digest(originalString.getBytes(StandardCharsets.UTF_8));
        return bytesToHex(encodedhash);
    }

    private static String bytesToHex(byte[] hash) {
        StringBuffer hexString = new StringBuffer();
        for (int i = 0; i < hash.length; i++) {
            String hex = Integer.toHexString(0xff & hash[i]);
            if(hex.length() == 1) hexString.append('0');
            hexString.append(hex);
        }
        return hexString.toString();
    }

    public static void sleep(){
        try {
            Thread.sleep(TimeUnit.MINUTES.toMillis(1));
        } catch (InterruptedException e) {
            log.error(e.toString());
        }
    }

    public static boolean shouldSleep() {
        return new Random().nextInt(10) == 1;
    }

    public static boolean shouldThrowTimeout() {
        return new Random().nextInt(10) == 1;
    }
}
